Changed bling worth to be an integer number of cents rather than a floating point dollar amount, to fix some off-by-1 errors.  It was originally a float because we could conceivably have 1 bling worth some fraction of cents, and keep track of fractions-of-fractions as integers would be a bit of a pain to do.  We could use BigDecimal, but it's not supported in GWT except by a side project that adds an extra 172kb of Javascript the user has to download.  So rather than following this rabbit hole to the end, we will just support bling worth in cents.

+++ src/java/com/threerings/msoy/money/data/all/BlingInfo.java
@@ -17,7 +17,7 @@ public class BlingInfo
     public int bling;
 
     /** The worth in US pennies. */
-    public float worthPerBling;
+    public int worthPerBling;
     
     /** The minimum amount of centibling required to cash out. */
     public int minCashOutBling;
@@ -26,7 +26,7 @@ public class BlingInfo
      * active. */
     public CashOutInfo cashOut;
     
-    public BlingInfo (int bling, float worthPerBling, int minCashOutBling, CashOutInfo cashOut)
+    public BlingInfo (int bling, int worthPerBling, int minCashOutBling, CashOutInfo cashOut)
     {
         this.bling = bling;
         this.worthPerBling = worthPerBling;
